pyproject.toml
[tool.poetry]
name = "fast-api-do-zero"
version = "0.1.0"
description = ""
authors = ["othon <othonbrener123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"

[tool.ruff]
line-length = 79
extend-exclude = ["migration"]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
